+++ src/components/Link/Link.tsx
@@ -33,6 +33,8 @@ export interface LinkProps {
   onClick?: (
     event: React.MouseEvent<HTMLDivElement | HTMLAnchorElement, MouseEvent>
   ) => void;
+  /** Prop that sets the HTML attribute to target where the link should go. */
+  target?: "_blank" | "_parent" | "_self" | "_top";
   /** Controls the link visuals: action, button, backwards, forwards, or default. */
   type?: LinkTypes;
 }
@@ -124,6 +126,7 @@ export const Link = chakra(
       href,
       id,
       onClick,
+      target,
       type = "default",
       ...rest
     } = props;
@@ -163,7 +166,8 @@ export const Link = chakra(
     }
     const styles = useMultiStyleConfig("Link", { variant });
     const rel = type === "external" ? "nofollow noopener noreferrer" : null;
-    const target = type === "external" ? "_blank" : null;
+    const internalTarget =
+      type === "external" ? "_blank" : target ? target : null;
     // Render with specific direction arrows if the type is
     // "forwards" or "backwards". Or render with the launch icon
     // if the type is "external". Otherwise, do not add an icon.
@@ -195,7 +199,7 @@ export const Link = chakra(
               ...childProps,
               ref,
               rel,
-              target,
+              target: internalTarget,
             },
             [childrenToClone.props.children]
           )}
@@ -209,7 +213,7 @@ export const Link = chakra(
           ref={ref}
           rel={rel}
           onClick={onClick}
-          target={target}
+          target={internalTarget}
           {...linkProps}
           __css={styles}
         >